Relocate generalError and serverError on especific folders
