+++ src/server/error/ServerError/types.ts
